Jetpack: Disable Resizing of Video Widgets, replay of r165722.
This was initially reverted by r165723, but then decided to roll back on because it fixes the problem and works in Jetpack as well. Summary: This diff seeks to fix the issues reported in https://[private link] and https://[private link] - where the jetpack-responsive-videos plugin wraps the output of core video widgets in a jetpack-video-wrapper div. The js logic in the plugin then resizes videos within these divs, which as of 4.9 causes some visual oddities in videos.

Test Plan: Using any of the themes that declare support for jetpack-responsive-videos add video widgets to the widget area. Then view the site and ensure the videos render and play properly. Also resize the window to see how things respond.

Reviewers: mermel

Differential Revision: https://[private link]

Merges r165801-wpcom.